# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
      "version": "1.0",
      "title": "API DE GERENCIAMENTO DE CONTATOS",
      "description": "Documentação da API criada em NodeJs e Typescript para o gerenciamento de contatos",
      "contact":{
        "name": "Suporte da API",
        "url": "https://www.ifpb.edu.br/cajazeiras"
      },
      "license": {
          "name": "IFPB - Campus Cajazeiras",
          "url": "https://www.ifpb.edu.br/cajazeiras"
      }
  },
  "servers": [
    {
      "url":"http://localhost:8080/",
      "description": "Servidor Local"
    }
  ],
  "paths":{
...
    "/contact/update":{
      "put":{
        "description": "Atualiza os dados de um contato da lista",
        "operationId": "contactUpdate",
        "parameters":[],
        "required": true,
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/ContactCreate"
              }
            }
          }
        },
        "responses": {
          "204":{
            "description": "RESPOSTA AO CONSEGUIR ATUALIZAR O CONTATO"
          },
          "500":{
            "description":"QUANDO UM ERRO OCORRER NA HORA DE SE ATUALIZAR UM CONTATO",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/MessageErrorContactUp"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components":{
    "schemas":{
...
      "telephone":{
        "type":"string",
        "example": "(83) 99647-1318"
      },
      "id":{
        "type": "number",
        "example": 1
      },
      "name":{
        "type":"string",
        "example": "Luciano Gouveia Rosa"
      },
...
      "ContactCreate":{
        "type": "object",
        "properties":{
          "id":{
            "$ref":"#/components/schemas/id"
          },
          "name":{
            "$ref":"#/components/schemas/name"
          },
          "telephone":{
            "$ref":"#/components/schemas/telephone"
          }
        }
      },
...
      "MessageErrorContactUpdate":{
        "type":"string",
        "example": "ERRO AO ATUALIZAR CONTATO"
      },
...
      "MessageErrorContactUp":{
        "type":"object",
        "properties":{
          "MENSAGEM":{
            "$ref":"#/components/schemas/MessageErrorContactUpdate"
          }
        }
      },
...
    }
  }
}